Interim fix for Pogie snapshots
Pogie sometimes calls F_READ with BC=0. In earlier versions of NextZXOS
this was incorrectly interpreted as 64K but has been fixed to mean 0
bytes (matching esxDOS) from v2.01B.

This is an interim fix for Pogie, which uses BC=0 but expects the
remainder of the file to be read. It replaces the following code:

        rst 8
        defb $9d ; F_READ
        ret nc

with a call to a new routine overwriting some debugger text, which does
the following:

        push af
        ld a,b
        or c
        jr nz,notzero
        dec bc
notzero:
        rst 8
        defb $9d ; F_READ
        ret c
        inc sp
        inc sp
        ret

In pogie.snx, the CALL is located at offset 0x473c and the routine at
offset 0x4163 in the snapshot.

In pogie2.snx, the CALL is located at offset 0x4792 and the routine at
offset 0x4263 in the snapshot.