Fix for intermittent 6ch-OnlyYou demo crashes
The OnlyYou demo performed 7ffd paging using the softcopy stored in the
system variable BANKM (0x5b5c). Unfortunately, in the .snx file this
byte was zero, instead of the usual 0x10, so paging with this value left
ROM2 paged instead of ROM3. Subsequent calls to the ROM therefore went
to random routines in ROM2.

Fixed the softcopy byte to 0x10.